digital-twin-service/requirements.txt
kafka-python
couchbase
fastapi==0.103.2
uvicorn[standard]==0.23.2